Edit: separação de APIs Graphql e Rest + Ajuste na geração do CSV com apenas 200 resutados + otimização na csv service para recuperar linhas e momento de onde parou

+++ Lab03/domain/generate_csv.py
@@ -1,24 +1,47 @@
-from Lab01.domain.repository_crawler_service import RepositoryCrawlerService
-from Lab01.domain.repository_model import Repository
-from Lab01.domain.repository_csv_service import RepositoryCsvService
+import os.path
+
+from Lab03.domain.repository_crawler_service import RepositoryCrawlerService
+from Lab03.domain.repository_model import Repository
+from Lab03.domain.repository_csv_service import RepositoryCsvService
+from settings import BASE_DIR
+
+CURSOR_PATH = f'{BASE_DIR}/Lab03/cursor.txt'
 
 
 def generate_repository_csv():
     repo_csv = RepositoryCsvService()
-    repo_csv.remove_file()
-    repo_csv.start_writer()
-    repo_csv.write_header()
-
     crawler = RepositoryCrawlerService()
+    if os.path.exists(CURSOR_PATH):
+        with open(CURSOR_PATH, 'r', encoding='utf-8') as f:
+            cursor = f.readline()
+        repo_csv.start_writer("r")
+        count = repo_csv.lines - 1
+    else:
+        cursor = ""
+        repo_csv.remove_file()
+        repo_csv.start_writer()
+        repo_csv.write_header()
+        count = 0
 
-    result = crawler.crawl()
-    for page in range(1, 11):
+    max_count = 200
+    result = crawler.crawl(cursor)
+    while True:
         for repo_json in result["repos"]:
             repo_model = Repository(repo_json)
             repo_csv.write_row(repo_model)
+            count += 1
+            if count == max_count:
+                break
+        if count == max_count:
+            break
+        f = open(CURSOR_PATH, 'w', encoding='utf-8')
+        f.write(f'{result["cursor"]}')
+        f.flush()
+        f.close()
         result = crawler.crawl(result["cursor"])
 
     repo_csv.reset_internal()
+    os.remove(CURSOR_PATH)
 
 
 def read_repository_csv():
